fix(config): preserve reversible route and BOM state
fix: upgrade Codex route journals to v7 while preserving native compaction and multi-agent feature choices across install, reconnect, migration, and uninstall.

fix: accept a single UTF-8 BOM at bridge, Codex TOML/journal, launcher state, runtime descriptor, manifest, updater, doctor, and tunnel manifest boundaries without changing each user file's BOM or line-ending convention.

refactor: split Codex integration, RuntimeHost, and RuntimeSupervisor responsibilities into focused modules capped below 500 lines while preserving their public CommonJS and TypeScript APIs.

test: cover v3-v6 journal recovery, exact route rollback, native passthrough, BOM round trips, runtime host transactions, supervisor lifecycle, and full root and Launcher regressions.
