+++ launcher/electron/json-file.cjs
@@ -1,0 +1,11 @@
+const fs = require("node:fs");
+
+function parseJsonText(text) {
+  return JSON.parse(text.replace(/^\uFEFF/, ""));
+}
+
+function readJsonFile(pathname) {
+  return parseJsonText(fs.readFileSync(pathname, "utf8"));
+}
+
+module.exports = { parseJsonText, readJsonFile };
