+++ src/codex-integration-journal.ts
@@ -1,0 +1,141 @@
+import { existsSync, readFileSync } from "node:fs";
+import { resolve } from "node:path";
+import { atomicWriteFile, stripUtf8Bom } from "./config";
+import {
+  getCodexConfigPath,
+  getCodexJournalPath,
+  getCodexJournalRecoveryPath,
+  serializeJournal,
+  writeFilesWithCompensation,
+} from "./codex-integration-shared";
+import type {
+  AnyCodexIntegrationJournal,
+  CodexIntegrationJournal,
+  LegacyCodexIntegrationJournal,
+  LegacyCodexIntegrationJournalV3,
+  LegacyCodexIntegrationJournalV4,
+  LegacyCodexIntegrationJournalV5,
+  LegacyCodexIntegrationJournalV6,
+} from "./codex-integration-shared";
+import { verifyManagedJournalState } from "./codex-integration-route";
+
+function parseJournal(path: string): AnyCodexIntegrationJournal {
+  const value = JSON.parse(stripUtf8Bom(readFileSync(path, "utf8"))) as Record<string, unknown>;
+  if (value.version === 7
+    && typeof value.active === "boolean"
+    && value.installed
+    && value.previous
+    && typeof value.configPath === "string") {
+    return value as unknown as CodexIntegrationJournal;
+  }
+  if (value.version === 6
+    && typeof value.active === "boolean"
+    && value.installed
+    && value.previous
+    && value.previousRemoteCompactionV2
+    && value.previousMultiAgent
+    && value.previousMultiAgentV2
+    && typeof value.configPath === "string") {
+    return value as unknown as LegacyCodexIntegrationJournalV6;
+  }
+  if (value.version === 5
+    && typeof value.active === "boolean"
+    && value.installed
+    && value.previous
+    && value.previousRemoteCompactionV2
+    && value.previousMultiAgent
+    && typeof value.configPath === "string") {
+    return value as unknown as LegacyCodexIntegrationJournalV5;
+  }
+  if (value.version === 4
+    && typeof value.active === "boolean"
+    && value.installed
+    && value.previous
+    && typeof value.configPath === "string") {
+    return value as unknown as LegacyCodexIntegrationJournalV4;
+  }
+  if (value.version === 3 && value.installed && value.previous && typeof value.configPath === "string") {
+    return value as unknown as LegacyCodexIntegrationJournalV3;
+  }
+  if (value.version === 2 && value.installed && value.previous && typeof value.providerBlock === "string") {
+    return value as unknown as LegacyCodexIntegrationJournal;
+  }
+  throw new Error(`Invalid Codex integration journal: ${path}`);
+}
+
+function journalMatchesConfig(journal: AnyCodexIntegrationJournal): boolean {
+  try {
+    assertJournalTargetsConfig(journal, getCodexConfigPath());
+    if (!existsSync(journal.configPath)) return false;
+    const text = readFileSync(journal.configPath, "utf8");
+    if (journal.version === 2) return text.includes(journal.providerBlock);
+    verifyManagedJournalState(text, journal);
+    return true;
+  } catch {
+    return false;
+  }
+}
+
+export function readJournal(): AnyCodexIntegrationJournal | undefined {
+  const primaryPath = getCodexJournalPath();
+  const recoveryPath = getCodexJournalRecoveryPath();
+  let primary: AnyCodexIntegrationJournal | undefined;
+  let recovery: AnyCodexIntegrationJournal | undefined;
+  let primaryError: unknown;
+  let recoveryError: unknown;
+  if (existsSync(primaryPath)) {
+    try { primary = parseJournal(primaryPath); } catch (error) { primaryError = error; }
+  }
+  if (existsSync(recoveryPath)) {
+    try { recovery = parseJournal(recoveryPath); } catch (error) { recoveryError = error; }
+  }
+  if (!primary && !recovery) {
+    if (primaryError) throw primaryError;
+    if (recoveryError) throw recoveryError;
+    return undefined;
+  }
+  if (primary && recovery && serializeJournal(primary) === serializeJournal(recovery)) return primary;
+  if (primary && !recovery && !recoveryError) {
+    atomicWriteFile(recoveryPath, serializeJournal(primary));
+    return primary;
+  }
+  if (recovery && !primary && !primaryError) {
+    if (!journalMatchesConfig(recovery)) {
+      throw new Error("Codex integration recovery journal does not match the active config");
+    }
+    atomicWriteFile(primaryPath, serializeJournal(recovery));
+    return recovery;
+  }
+
+  const primaryMatches = primary ? journalMatchesConfig(primary) : false;
+  const recoveryMatches = recovery ? journalMatchesConfig(recovery) : false;
+  if (primaryMatches === recoveryMatches) {
+    throw new Error(
+      primaryMatches
+        ? "Codex integration journal copies contain different baselines for the same config"
+        : "Codex integration journal copies do not match the active config",
+    );
+  }
+  const selected = primaryMatches ? primary! : recovery!;
+  const data = serializeJournal(selected);
+  writeFilesWithCompensation([
+    { path: recoveryPath, data },
+    { path: primaryPath, data },
+  ]);
+  return selected;
+}
+
+export function assertJournalTargetsConfig(
+  journal: AnyCodexIntegrationJournal,
+  configPath: string,
+): void {
+  const pathIdentity = (value: string): string => {
+    const normalized = resolve(value);
+    return process.platform === "win32" ? normalized.toLowerCase() : normalized;
+  };
+  if (pathIdentity(journal.configPath) !== pathIdentity(configPath)) {
+    throw new Error(
+      `Codex integration journal belongs to ${journal.configPath}, not the active config ${configPath}`,
+    );
+  }
+}
