+++ src/codex-integration-shared.ts
@@ -1,0 +1,247 @@
+import { createHash } from "node:crypto";
+import { existsSync, mkdirSync, readFileSync, rmSync } from "node:fs";
+import { homedir } from "node:os";
+import { dirname, join, resolve } from "node:path";
+import type { AppConfig } from "./config";
+import { atomicWriteFile, expandUserPath, getConfigDir } from "./config";
+
+export const MANAGED_COMMENT = "# Managed by codex-chatgpt-web; `codex-chatgpt-web uninstall` restores prior values.";
+export const MANAGED_REMOTE_COMPACTION_LINE =
+  "remote_compaction_v2 = false # Managed by codex-chatgpt-web: bounds retained Web image history.";
+export const MANAGED_MULTI_AGENT_LINE =
+  "multi_agent = true # Managed by codex-chatgpt-web: enables routed Web subagents.";
+export const MANAGED_MULTI_AGENT_V2_LINE =
+  "multi_agent_v2 = false # Managed by codex-chatgpt-web: keeps routed Web subagent payloads readable.";
+export const MANAGED_MULTI_AGENT_V2_TABLE_LINE =
+  "enabled = false # Managed by codex-chatgpt-web: keeps routed Web subagent payloads readable.";
+
+export interface PreviousAssignment {
+  present: boolean;
+  rawLine?: string;
+  value?: string;
+  index?: number;
+}
+
+export type ManagedAssignmentKey = "openai_base_url" | "model_provider" | "model_catalog_json";
+
+export interface PreviousFeatureAssignment extends PreviousAssignment {
+  tablePresent: boolean;
+  tableName?: "features" | "features.multi_agent_v2";
+}
+
+export interface CodexIntegrationJournal {
+  version: 7;
+  active: boolean;
+  configPath: string;
+  installed: {
+    openai_base_url: string;
+  };
+  previous: Record<ManagedAssignmentKey, PreviousAssignment>;
+  format?: {
+    lineEnding: "\n" | "\r\n";
+    trailingNewline: boolean;
+  };
+}
+
+export interface LegacyCodexIntegrationJournalV6 {
+  version: 6;
+  active: boolean;
+  configPath: string;
+  installed: {
+    openai_base_url: string;
+    remote_compaction_v2: false;
+    multi_agent: true;
+    multi_agent_v2: false;
+  };
+  previous: Record<ManagedAssignmentKey, PreviousAssignment>;
+  previousRemoteCompactionV2: PreviousFeatureAssignment;
+  previousMultiAgent: PreviousFeatureAssignment;
+  previousMultiAgentV2: PreviousFeatureAssignment;
+  format?: {
+    lineEnding: "\n" | "\r\n";
+    trailingNewline: boolean;
+  };
+}
+
+export interface LegacyCodexIntegrationJournalV5 {
+  version: 5;
+  active: boolean;
+  configPath: string;
+  installed: {
+    openai_base_url: string;
+    remote_compaction_v2: false;
+    multi_agent: true;
+  };
+  previous: Record<ManagedAssignmentKey, PreviousAssignment>;
+  previousRemoteCompactionV2: PreviousFeatureAssignment;
+  previousMultiAgent: PreviousFeatureAssignment;
+  format?: {
+    lineEnding: "\n" | "\r\n";
+    trailingNewline: boolean;
+  };
+}
+
+export interface LegacyCodexIntegrationJournalV4 {
+  version: 4;
+  active: boolean;
+  configPath: string;
+  installed: {
+    openai_base_url: string;
+  };
+  previous: Record<ManagedAssignmentKey, PreviousAssignment>;
+  format?: {
+    lineEnding: "\n" | "\r\n";
+    trailingNewline: boolean;
+  };
+}
+
+export interface LegacyCodexIntegrationJournalV3 {
+  version: 3;
+  configPath: string;
+  installed: {
+    openai_base_url: string;
+  };
+  previous: Record<ManagedAssignmentKey, PreviousAssignment>;
+  format?: {
+    lineEnding: "\n" | "\r\n";
+    trailingNewline: boolean;
+  };
+}
+
+export interface LegacyCodexIntegrationJournal {
+  version: 2;
+  configPath: string;
+  catalogPath: string;
+  catalogSha256: string;
+  providerBlock: string;
+  installed: {
+    model_provider: string;
+    model_catalog_json: string;
+  };
+  previous: {
+    model_provider: PreviousAssignment;
+    model_catalog_json: PreviousAssignment;
+  };
+}
+
+export type ManagedRouteJournal =
+  | CodexIntegrationJournal
+  | LegacyCodexIntegrationJournalV6
+  | LegacyCodexIntegrationJournalV5
+  | LegacyCodexIntegrationJournalV4
+  | LegacyCodexIntegrationJournalV3;
+export type AnyCodexIntegrationJournal = ManagedRouteJournal | LegacyCodexIntegrationJournal;
+
+export interface FileSnapshot {
+  path: string;
+  exists: boolean;
+  data?: Buffer;
+}
+
+export interface InstallCodexIntegrationOptions {
+  replaceExistingRoute?: boolean;
+}
+
+export interface UninstallCodexIntegrationResult {
+  changed: boolean;
+}
+
+export interface SetCodexIntegrationActiveResult {
+  changed: boolean;
+  active: boolean;
+}
+
+export interface CodexModelContextOverride {
+  model: string;
+  contextWindow: number;
+}
+
+export function getCodexHome(): string {
+  const configured = process.env.CODEX_HOME?.trim();
+  return resolve(expandUserPath(configured || join(homedir(), ".codex")));
+}
+
+export function getCodexConfigPath(): string {
+  return join(getCodexHome(), "config.toml");
+}
+
+export function getCodexModelsCachePath(): string {
+  return join(getCodexHome(), "models_cache.json");
+}
+
+export function getCodexJournalPath(): string {
+  return join(getConfigDir(), "codex", "integration-journal.json");
+}
+
+export function getCodexJournalRecoveryPath(): string {
+  return join(getConfigDir(), "codex", "integration-journal.recovery.json");
+}
+
+export function routeUrl(config: AppConfig): string {
+  return `http://${config.host}:${config.port}/v1`;
+}
+
+export function sha256(value: string | Uint8Array): string {
+  return createHash("sha256").update(value).digest("hex");
+}
+
+export function snapshotFile(path: string): FileSnapshot {
+  return existsSync(path)
+    ? { path, exists: true, data: readFileSync(path) }
+    : { path, exists: false };
+}
+
+export function restoreFileSnapshot(snapshot: FileSnapshot): void {
+  if (snapshot.exists) {
+    if (!snapshot.data) throw new Error(`File snapshot is missing data: ${snapshot.path}`);
+    atomicWriteFile(snapshot.path, snapshot.data);
+  } else {
+    rmSync(snapshot.path, { force: true });
+  }
+}
+
+export function writeFilesWithCompensation(
+  writes: Array<{ path: string; data: string | Uint8Array }>,
+  removals: string[] = [],
+): void {
+  const paths = [...new Set([...writes.map(write => write.path), ...removals])];
+  const snapshots = paths.map(snapshotFile);
+  try {
+    for (const write of writes) atomicWriteFile(write.path, write.data);
+    for (const removal of removals) rmSync(removal, { force: true });
+  } catch (error) {
+    const rollbackFailures: string[] = [];
+    for (const snapshot of [...snapshots].reverse()) {
+      try {
+        restoreFileSnapshot(snapshot);
+      } catch (rollbackError) {
+        rollbackFailures.push(`${snapshot.path}: ${rollbackError instanceof Error ? rollbackError.message : String(rollbackError)}`);
+      }
+    }
+    const primary = error instanceof Error ? error.message : String(error);
+    throw new Error(
+      rollbackFailures.length > 0
+        ? `${primary}; Codex integration rollback also failed: ${rollbackFailures.join("; ")}`
+        : primary,
+    );
+  }
+}
+
+export function serializeJournal(journal: AnyCodexIntegrationJournal): string {
+  return `${JSON.stringify(journal, null, 2)}\n`;
+}
+
+export function writeIntegrationState(
+  journal: AnyCodexIntegrationJournal,
+  configWrite?: { path: string; data: string },
+  removals: string[] = [],
+): void {
+  const data = serializeJournal(journal);
+  // The recovery copy records intent and the primary copy records commit. If the process stops
+  // between those writes, the physical config unambiguously selects the completed state.
+  writeFilesWithCompensation([
+    { path: getCodexJournalRecoveryPath(), data },
+    ...(configWrite ? [configWrite] : []),
+    { path: getCodexJournalPath(), data },
+  ], removals);
+}
